Change timestamp format for new change files
This changes the current format for 'codelog new'-generated
change files from Unix Epoch format to a human-readable format.

former: 1516389093_change.yml
new: 20180119171207818_change.yml

The rationale behid it is that when searching for a partial
changelog file between releases is much easier if we can search
by filename. For example: I know a change was made yesterday during
the morning, so I can 'easily' know which file corresponds to it,
even if we have a lot of yml files under 'unreleased' directory.

Signed-off-by: Luiz Cavalcanti <luiz.cavalcanti@codus.com.br>